Moved DB engine call to derived class, renamed db config
